Broke out ChatServer.js into module; preparation for adding E2E testing
