+++ lib/ChatRoom.js
